--- src/index.js
@@ -1,2 +1,0 @@
-import './style.css';
-import './modules/main.js';
