--- src/modules/saveTasksToLocalStorage.js
@@ -1,7 +1,0 @@
-import List from './list.js';
-
-const saveTasksToLocalStorage = () => {
-  localStorage.setItem('tasks', JSON.stringify(List.taskList));
-};
-
-export default saveTasksToLocalStorage;
